+++ docs/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+## 8.38.0 <span class="changelog-date">(2024-09-22)</span>
+
 * Renamed `is:class` to `is:subclass`.
 * Fixed D1 item category searches such as `is:primary` and `is:horn`. Now they're up to parity with the D2 categories.